feature products is reacting...
