+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/alamin/Documents/build-it-best/src/App.js":"1","/home/alamin/Documents/build-it-best/src/components/product/ProductList.js":"2","/home/alamin/Documents/build-it-best/src/index.js":"3","/home/alamin/Documents/build-it-best/src/components/Slider.js":"4"},{"size":443,"mtime":1611415171434,"results":"5","hashOfConfig":"6"},{"size":2302,"mtime":1611415805165,"results":"7","hashOfConfig":"6"},{"size":478,"mtime":1611415120359,"results":"8","hashOfConfig":"6"},{"size":2434,"mtime":1611413604825,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"kc5qlf",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/alamin/Documents/build-it-best/src/App.js",[],"/home/alamin/Documents/build-it-best/src/components/product/ProductList.js",[],"/home/alamin/Documents/build-it-best/src/index.js",[],"/home/alamin/Documents/build-it-best/src/components/Slider.js",[]]
+[{"/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/components/product/ProductList.js":"1","/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/App.js":"2","/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/components/featured-products/FeatureProducts.js":"3","/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/components/featured-products/FeatureImage.js":"4","/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/components/product/ProductCard.js":"5","/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/components/Slider.js":"6"},{"size":2041,"mtime":1611418532317,"results":"7","hashOfConfig":"8"},{"size":452,"mtime":1611889181506,"results":"9","hashOfConfig":"8"},{"size":2669,"mtime":1611928371757,"results":"10","hashOfConfig":"8"},{"size":226,"mtime":1611927586116,"results":"11","hashOfConfig":"8"},{"size":352,"mtime":1611418532316,"results":"12","hashOfConfig":"8"},{"size":2434,"mtime":1611165444915,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"97abac",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":"20"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/components/product/ProductList.js",[],"/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/App.js",[],"/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/components/featured-products/FeatureProducts.js",["27","28"],"import { ArrowRightOutlined } from \"@ant-design/icons\";\nimport { Row, Col, Image, Typography } from \"antd\";\nimport styles from \"./Feature.module.css\";\nimport FeatureImage from \"./FeatureImage\";\n\nimport watchImage from \"../../assets/images/watch.jpg\";\nimport glassImage from \"../../assets/images/glass.jpg\";\nimport bagpackImage from \"../../assets/images/bagpack.jpg\";\n\nconst { Link } = Typography;\n\nconst FeatureProducts = () => {\n  return (\n    <>\n      <div className={`${styles.bgColor} ${styles.block}`}>\n        <div className={styles.containerFluid}>\n          <div className={styles.titleHolder}>\n            <h2>Feature Products</h2>\n          </div>\n          <Row>\n            <Col span={16}>\n              {/* Image of Watch */}\n              <Row>\n                <div className={`${styles.container} ${styles.image}`}>\n                  <Link>\n                    <FeatureImage\n                      source={watchImage}\n                      style={{\n                        marginBottom: \"10px\",\n                        border: \"1px solid black\",\n                        objectFit: \"cover\",\n                      }}\n                    />\n                  </Link>                  \n                  {/*<div className=\"overlay\">\n                    <Link href=\"#\" target=\"_blank\" style={{ color: \"white\" }}>\n                      Watch\n                      <ArrowRightOutlined />\n                    </Link>\n                  </div>*/}\n                </div>\n              </Row>\n              {/* Image of Glass */}\n              <Row>\n                <div className={`${styles.container} ${styles.image}`}>\n                  <FeatureImage\n                    source={glassImage}\n                    style={{ \n                      border: \"1px solid black\", \n                      objectFit: \"cover\",\n                    }}\n                  />\n                  {/* <div className=\"overlayGlass\">\n                    <Link href=\"#\" target=\"_blank\" style={{ color: \"white\" }}>\n                      Glass\n                      <ArrowRightOutlined />\n                    </Link>\n                  </div> */}\n                </div>\n              </Row>\n            </Col>\n            <Col span={8}>\n              {/* Image of Bagpack */}\n              <FeatureImage \n                  source = {bagpackImage}\n                  style={{\n                    margin: \"0 0 0 15px\",\n                    border: \"1px solid black\",\n                    objectFit: \"cover\",\n                    height: \"1050px\",\n                  }}\n              />\n            </Col>\n          </Row>\n        </div>\n      </div>\n    </>\n  );\n};\n\nexport default FeatureProducts;\n","/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/components/featured-products/FeatureImage.js",[],"/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/components/product/ProductCard.js",[],"/media/el06/Coding/Projects/2. Practice Projects/4. Pluto/pluto/src/components/Slider.js",[],{"ruleId":"29","severity":1,"message":"30","line":1,"column":10,"nodeType":"31","messageId":"32","endLine":1,"endColumn":28},{"ruleId":"29","severity":1,"message":"33","line":2,"column":20,"nodeType":"31","messageId":"32","endLine":2,"endColumn":25},"no-unused-vars","'ArrowRightOutlined' is defined but never used.","Identifier","unusedVar","'Image' is defined but never used."]
