+++ src/components/featured-products/FeatureImage.js
@@ -1,0 +1,14 @@
+import { Image } from "antd";
+const FeatureImage = ({source, style}) => {
+  return(
+    <>
+      <Image 
+        src={source}
+        preview={false}
+        style={style}
+      />
+    </>
+  );
+};
+
+export default FeatureImage;
